+++ src/components/App/Enum.ts
